requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-extensions==3.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
